refactor(lib): split math.c into math_sqrt.c + math.c
The B6 chantier wired `_DEP_hdma := dma math` in make/common.mk so
hdma's iris-wipe could call sqrt16 directly. That dragged the full
math module — sine_table (512 B), atan_lut (65 B), atan2_8 (~480 B),
fixSin/fixCos/fixDiv/fixMul/fixAbs/fixClamp/fixLerp, mul16/div16/mod16
— into every hdma-using example, even when the user only needed sqrt
through hdma. Six examples felt that bloat with a measurable
bank-ROM cost.

Split shipped as a follow-up clean-up:

  - lib/source/math_sqrt.c (new) — sqrt16 + fixSqrt only.
  - lib/source/math.c (existing) — everything else (sine LUT, atan
    LUT, fixSin/fixCos, atan2_8, arithmetic helpers).
  - make/common.mk — `_DEP_hdma := dma math_sqrt` (was: dma math),
    `_DEP_math := math_sqrt` so user code that lists `math` still
    transitively pulls sqrt for free; the resolver flattens the chain.

The header `<snes/math.h>` is unchanged — it declares both halves and
clients see one API surface.

Concurrent clean-up: 5 examples had `math` listed in LIB_MODULES but
never called any math function. Removing the dead listing (window,
parallax_scrolling, hdma_gradient, transparent_window, gradient_colors)
makes their LIB_MODULES match what their main.c actually uses.

Bank-$00 free space on hdma-using examples (post-build):

  Example              before   after   delta
  ─────────────────────────────────────────────
  hdma_wave            10758    12479    +1721
  gradient_colors      12324    14045    +1721
  parallax_scrolling    5628     7349    +1721
  transparent_window    8084     9805    +1721
  hdma_gradient         9349    11070    +1721
  hdma_helpers          7865     9586    +1721
  window                  66       66       0

The ~1721 B improvement on six examples is the math module's
runtime + sine_table footprint (the constant-folded delta is the same
because `_DEP_hdma` change is identical for every hdma user).

The window result (unchanged at 66 B free) is informative: window's
bank-$00 cliff is graphics-data-bound, not math-bound. Even with the
math bloat removed, the linker's SUPERFREE placement filled the freed
bank-$00 slots with content that was previously spilling to bank-$01.
The split still benefits window indirectly — bank-$01's free space
grows by ~1721 B, which gives users adding new const data more
runway before bank-$00 has to absorb it.

Validation: full test suite returns 409/409 PASS (no `--allow-known-bugs`).
Bank-$00 ratchet at threshold 16 still satisfied for every example.
`<snes/math.h>` API unchanged; user-visible API stable.

Catalogue cross-reference: math_sqrt.c's docstring documents the
split rationale and refers back to .claude/STRUCTURAL_DEFECTS.md
chantier B6 for the original sqrt+atan addition.
