+++ examples/graphics/effects/gradient_colors/Makefile
@@ -6,7 +6,7 @@ TARGET   := gradient_colors.sfc
 ROM_NAME := OPENSNES GRADCOLORS
 
 USE_LIB     := 1
-LIB_MODULES := console dma background sprite hdma input math
+LIB_MODULES := console dma background sprite hdma input
 
 CSRC   := main.c
 ASMSRC := data.asm
