+++ examples/graphics/effects/hdma_gradient/Makefile
@@ -6,7 +6,7 @@ TARGET   := hdma_gradient.sfc
 ROM_NAME := HDMA GRADIENT DEMO
 
 USE_LIB     := 1
-LIB_MODULES := console dma background sprite hdma input math
+LIB_MODULES := console dma background sprite hdma input
 
 CSRC   := main.c
 ASMSRC := data.asm
