+++ examples/graphics/effects/transparent_window/Makefile
@@ -1,6 +1,6 @@
 #==============================================================================
 # Transparent Window Example
-# Ported from PVSnesLib — semi-transparent rectangle using color math + HDMA
+# Ported from PVSnesLib — semi-transparent rectangle using color + HDMA
 #==============================================================================
 
 OPENSNES := $(shell cd ../../../.. && pwd)
@@ -10,7 +10,7 @@ ROM_NAME   := TRANSPARENT WINDOW
 
 # Use the OpenSNES library
 USE_LIB    := 1
-LIB_MODULES := console sprite dma background window colormath hdma math
+LIB_MODULES := console sprite dma background window colormath hdma
 
 # Source files
 CSRC       := main.c
